requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
simplejson==3.19.2
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
